aso_project_CirciuMihneaTeodor_faza3/project/requirements.txt
djangorestframework
Django==4.1.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
